+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `BuilderConnectorWidgetsOnly` - a degraded version of `BuilderConnector` for
+  getting widgets after the signals were connected.
 
 ## 0.3.0 - 2021-02-09
 ### Added